# back-end/app/Http/Requests/CategoryRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Contracts\Validation\ValidationRule;
use Illuminate\Foundation\Http\FormRequest;
use OpenApi\Attributes as OA;

#[OA\Schema(
	schema: 'CategoryRequest',
	required: ['name'],
	properties: [
		new OA\Property(property: 'name', type: 'string', maxLength: 255, example: 'Haircut'),
	]
)]
class CategoryRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
		return auth()->user()?->isAdmin() ?? false;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, ValidationRule|array|string>
     */
    public function rules(): array
    {
        return [
			'name' => 'required|string|max:255',
        ];
    }
}